requirements.txt
# Core dependencies
supabase>=2.0.0
pandas>=2.1.0
numpy>=1.25.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7

# API and networking
requests==2.31.0
httpx==0.24.1

# Data processing
openpyxl==3.1.2
pyyaml==6.0.1
ujson==5.8.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1

# Utilities
tqdm==4.66.1
rich==13.5.2